カレンダーイベントハンドラーをリファクタリングし、処理をCalendarEventsUseCaseクラスに移動しました。新たにCalendarEventsUseCaseを実装し、GoogleカレンダーAPIとの連携を整理しました。
