+++ src/handlers/calendar-events-handler.ts
@@ -1,11 +1,8 @@
-import { CalendarEventsNotifier } from "../calendar-events-notifier";
-import { GoogleCalendarApiAdapter } from "../google-calendar-api-adapter";
-import { GoogleAuthAdapter } from "../lib/google-auth-adapter";
-import { Config } from "../lib/config";
-import { AwsParameterFetcher } from "../lib/aws-parameter-fetcher";
-import { LineMessagingApiClient } from "../line-messaging-api-client";
 import { ApiResponseBuilder } from "../lib/api-response-builder";
 import type { APIGatewayProxyResult } from "aws-lambda";
+import { CalendarEventsUseCase } from "../usecases/calendar-events-usecase";
+import { Config } from "../lib/config";
+import { AwsParameterFetcher } from "../lib/aws-parameter-fetcher";
 
 export const calendarEventsHandler =
   async (): Promise<APIGatewayProxyResult> => {
@@ -16,19 +13,7 @@ export const calendarEventsHandler =
       const parameterFetcher = new AwsParameterFetcher();
       await config.init(parameterFetcher);
 
-      const auth = new GoogleAuthAdapter();
-      const token = {
-        accessToken: config.GOOGLE_ACCESS_TOKEN,
-        refreshToken: config.GOOGLE_REFRESH_TOKEN,
-      };
-      auth.setTokens(token);
-      const googleCalendarApi = new GoogleCalendarApiAdapter(auth);
-      const lineMessagingApiClient = new LineMessagingApiClient();
-
-      await new CalendarEventsNotifier(
-        googleCalendarApi,
-        lineMessagingApiClient
-      ).call();
+      await new CalendarEventsUseCase().execute();
       console.info("End calendar events handler");
 
       return responseBuilder.success(
